+++ changelog.md
@@ -11,7 +11,8 @@
 * Make `IO` benchmarks immune to `-fspec-constr-count` limit.
 * Decomission `debug` build flag.
 * Decomission warning when `--timeout` is absent.
-* Add `instance Eq RelStDev` and `instance Ord RelStDev`.
+* Add `instance {Eq,Ord,Num,Fractional} {RelStDev,FailIfSlower,FailIfFaster}`.
+* Add `instance {Eq,Ord} {CsvPath,SvgPath,BaselinePath}`.
 * Add `unRelStDev :: RelStDev -> Double`.
 
 # 0.3.5